refactoring spectral profile plots included spectral processing dialog
